Implemented log syncing optimization.
When a follower returns log inconsistency results the leader must
retry syncing with the previous log index. This optimization skips the
syncing to the beginning of the first index of the most recent term on
the follower. This can make things quite fast for when the follower is
quite a bit behind.